feat : ayheon
feat

+++ src/main/java/com/ot/main/delivery/data/repository/DeliveryRepository.java
@@ -4,6 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ot.main.delivery.data.entity.Delivery;
 
-public interface DeliveryRepository extends JpaRepository<Delivery, Integer> {
+public interface DeliveryRepository extends JpaRepository<Delivery, Long> {
+
 
 }
